Done with backend

+++ sinatra/app/controllers/applications_controller.rb
@@ -35,14 +35,20 @@ class ApplicationsController < ApplicationController
   end
 
   delete '/applications/delete/:id' do
-    application = Application.find_by(id: params[:id])
-    if application
-      application.destroy
-      message = { success: "Application has been deleted successfully" }
+    if session[:user_id]
+      application = Application.find_by(id: params[:id])
+      if application
+        application.destroy
+        message = { success: "Application has been deleted successfully" }
+      else
+        status 404
+        message = { error: "Application not found" }
+      end
     else
-      status 404
-      message = { error: "Application not found" }
+      status 401
+      message = { error: "Not logged in" }
     end
+
     message.to_json()
   end
 end
